Enum crc_any::CRC

pub enum CRC {
    CRCu8(CRCu8),
    CRCu16(CRCu16),
    CRCu32(CRCu32),
    CRCu64(CRCu64),
}

This struct can help you compute a CRC value.

Variants

CRCu8(CRCu8)

CRCu16(CRCu16)

CRCu32(CRCu32)

CRCu64(CRCu64)

Methods

impl CRC

pub fn create_crc(
poly: u64,
bits: u8,
initial: u64,
final_xor: u64,
reflect: bool
) -> CRC

Create a CRC instance by providing the length of bits, expression, reflection, an initial value and a final xor value.

pub fn create_crc_u8(
poly: u8,
bits: u8,
initial: u8,
final_xor: u8,
reflect: bool
) -> CRC

Create a CRC instance by providing the length of bits, expression, reflection, an initial value and a final xor value.

pub fn create_crc_u16(
poly: u16,
bits: u8,
initial: u16,
final_xor: u16,
reflect: bool
) -> CRC

Create a CRC instance by providing the length of bits, expression, reflection, an initial value and a final xor value.

pub fn create_crc_u32(
poly: u32,
bits: u8,
initial: u32,
final_xor: u32,
reflect: bool
) -> CRC

Create a CRC instance by providing the length of bits, expression, reflection, an initial value and a final xor value.

pub fn create_crc_u64(
poly: u64,
bits: u8,
initial: u64,
final_xor: u64,
reflect: bool
) -> CRC

Create a CRC instance by providing the length of bits, expression, reflection, an initial value and a final xor value.

pub fn digest<T: ?Sized + AsRef<[u8]>>(&mut self, data: &T)

Digest some data.

pub fn reset(&mut self)

Reset the sum.

pub fn get_crc(&mut self) -> u64

Get the current CRC value (it always returns a u64 value). You can continue calling digest method even after getting a CRC value.

impl CRC

pub fn get_crc_vec_le(&mut self) -> Vec<u8>

Get the current CRC value (it always returns a vec instance with a length corresponding to the CRC bits). You can continue calling digest method even after getting a CRC value.

pub fn get_crc_vec_be(&mut self) -> Vec<u8>

Get the current CRC value (it always returns a vec instance with a length corresponding to the CRC bits). You can continue calling digest method even after getting a CRC value.

impl CRC

pub fn crc3gsm() -> CRC

pub fn crc4itu() -> CRC

pub fn crc4interlaken() -> CRC

pub fn crc5epc() -> CRC

pub fn crc5itu() -> CRC

pub fn crc5usb() -> CRC

pub fn crc6cdma2000_a() -> CRC

pub fn crc6cdma2000_b() -> CRC

pub fn crc6darc() -> CRC

pub fn crc6gsm() -> CRC

pub fn crc6itu() -> CRC

pub fn crc7() -> CRC

pub fn crc7umts() -> CRC

pub fn crc8() -> CRC

pub fn crc8cdma2000() -> CRC

pub fn crc8darc() -> CRC

pub fn crc8dvb_s2() -> CRC

pub fn crc8ebu() -> CRC

pub fn crc8icode() -> CRC

pub fn crc8itu() -> CRC

pub fn crc8maxim() -> CRC

pub fn crc8rohc() -> CRC

pub fn crc8wcdma() -> CRC

pub fn crc10() -> CRC

pub fn crc10cdma2000() -> CRC

pub fn crc10gsm() -> CRC

pub fn crc11() -> CRC

pub fn crc12() -> CRC

pub fn crc12cdma2000() -> CRC

pub fn crc12gsm() -> CRC

pub fn crc13bbc() -> CRC

pub fn crc14darc() -> CRC

pub fn crc14gsm() -> CRC

pub fn crc15can() -> CRC

pub fn crc15mpt1327() -> CRC

pub fn crc16() -> CRC

pub fn crc16ccitt_false() -> CRC

pub fn crc16aug_ccitt() -> CRC

pub fn crc16buypass() -> CRC

pub fn crc16cdma2000() -> CRC

pub fn crc16dds_110() -> CRC

pub fn crc16dect_r() -> CRC

pub fn crc16dect_x() -> CRC

pub fn crc16dnp() -> CRC

pub fn crc16en_13757() -> CRC

pub fn crc16genibus() -> CRC

pub fn crc16maxim() -> CRC

pub fn crc16mcrf4cc() -> CRC

pub fn crc16riello() -> CRC

pub fn crc16t10_dif() -> CRC

pub fn crc16teledisk() -> CRC

pub fn crc16tms13157() -> CRC

pub fn crc16usb() -> CRC

pub fn crc_a() -> CRC

pub fn crc16kermit() -> CRC

pub fn crc16modbus() -> CRC

pub fn crc16_x25() -> CRC

pub fn crc16xmodem() -> CRC

pub fn crc17can() -> CRC

pub fn crc21can() -> CRC

pub fn crc24() -> CRC

pub fn crc24ble() -> CRC

pub fn crc24flexray_a() -> CRC

pub fn crc24flexray_b() -> CRC

pub fn crc24lte_a() -> CRC

pub fn crc24lte_b() -> CRC

pub fn crc24os9() -> CRC

pub fn crc30cdma() -> CRC

pub fn crc32() -> CRC

pub fn crc32mhash() -> CRC

pub fn crc32bzip2() -> CRC

pub fn crc32c() -> CRC

pub fn crc32d() -> CRC

pub fn crc32mpeg2() -> CRC

pub fn crc32posix() -> CRC

pub fn crc32q() -> CRC

pub fn crc32jamcrc() -> CRC

pub fn crc32xfer() -> CRC

pub fn crc40gsm() -> CRC

pub fn crc64() -> CRC

pub fn crc64iso() -> CRC

pub fn crc64jones() -> CRC