Struct crc_any::CRCu8

pub struct CRCu8 { /* fields omitted */ }

This struct can help you compute a CRC-8 (or CRC-x where x is under 8) value.

Methods

impl CRCu8

pub fn create_crc(
poly: u8,
bits: u8,
initial: u8,
final_xor: u8,
reflect: bool
) -> CRCu8

Create a CRCu8 instance by providing the length of bits, expression, reflection, an initial value and a final xor value.

pub fn digest<T: ?Sized + AsRef<[u8]>>(&mut self, data: &T)

Digest some data.

pub fn reset(&mut self)

Reset the sum.

pub fn get_crc(&self) -> u8

Get the current CRC value (it always returns a u8 value). You can continue calling digest method even after getting a CRC value.

impl CRCu8

pub fn crc3gsm() -> CRCu8

pub fn crc4itu() -> CRCu8

pub fn crc4interlaken() -> CRCu8

pub fn crc5epc() -> CRCu8

pub fn crc5itu() -> CRCu8

pub fn crc5usb() -> CRCu8

pub fn crc6cdma2000_a() -> CRCu8

pub fn crc6cdma2000_b() -> CRCu8

pub fn crc6darc() -> CRCu8

pub fn crc6gsm() -> CRCu8

pub fn crc6itu() -> CRCu8

pub fn crc7() -> CRCu8

pub fn crc7umts() -> CRCu8

pub fn crc8() -> CRCu8

pub fn crc8cdma2000() -> CRCu8

pub fn crc8darc() -> CRCu8

pub fn crc8dvb_s2() -> CRCu8

pub fn crc8ebu() -> CRCu8

pub fn crc8icode() -> CRCu8

pub fn crc8itu() -> CRCu8

pub fn crc8maxim() -> CRCu8

pub fn crc8rohc() -> CRCu8

pub fn crc8wcdma() -> CRCu8