crc_any

Struct CRCu64

pub struct CRCu64 { /* private fields */ }

This struct can help you compute a CRC-64 (or CRC-x where x is equal or less than 64) value.

Implementations

impl CRCu64

pub fn create_crc( poly: u64, bits: u8, initial: u64, final_xor: u64, reflect: bool, ) -> CRCu64

Create a CRCu64 instance by providing the length of bits, expression, reflection, an initial value and a final xor value.

pub fn digest<T: ?Sized + AsRef<[u8]>>(&mut self, data: &T)

Digest some data.

pub fn reset(&mut self)

Reset the sum.

pub fn get_crc(&self) -> u64

Get the current CRC value (it always returns a u64 value). You can continue calling digest method even after getting a CRC value.

impl CRCu64

pub fn crc40gsm() -> CRCu64

Check	Poly	Init	Ref	XorOut

0xD4164FC646	0x0004820009	0x0000000000	false	0xFFFFFFFFFF

let mut crc = CRCu64::crc40gsm();
crc.digest(b"123456789");

pub fn crc64() -> CRCu64

Check	Poly	Init	Ref	XorOut

0x6C40DF5F0B497347	0x42F0E1EBA9EA3693	0x0000000000000000	false	0x0000000000000000

let mut crc = CRCu64::crc64();
crc.digest(b"123456789");

pub fn crc64iso() -> CRCu64

Check	Poly	Init	Ref	XorOut

0xB90956C775A41001	0x000000000000001B (rev: 0xD800000000000000)	0xFFFFFFFFFFFFFFFF	true	0xFFFFFFFFFFFFFFFF

let mut crc = CRCu64::crc64iso();
crc.digest(b"123456789");

pub fn crc64we() -> CRCu64

Check	Poly	Init	Ref	XorOut

0x62EC59E3F1A4F00A	0x42F0E1EBA9EA3693	0xFFFFFFFFFFFFFFFF	false	0xFFFFFFFFFFFFFFFF

let mut crc = CRCu64::crc64we();
crc.digest(b"123456789");

pub fn crc64jones() -> CRCu64

Check	Poly	Init	Ref	XorOut

0xE9C6D914C4B8D9CA	0xAD93D23594C935A9 (rev: 0x95AC9329AC4BC9B5)	0x0000000000000000	true	0x0000000000000000

let mut crc = CRCu64::crc64jones();
crc.digest(b"123456789");