crc_any

Struct CRCu8

pub struct CRCu8 { /* private fields */ }

This struct can help you compute a CRC-8 (or CRC-x where x is equal or less than 8) value.

Implementations

impl CRCu8

pub fn create_crc( poly: u8, bits: u8, initial: u8, final_xor: u8, reflect: bool, ) -> CRCu8

Create a CRCu8 instance by providing the length of bits, expression, reflection, an initial value and a final xor value.

pub fn digest<T: ?Sized + AsRef<[u8]>>(&mut self, data: &T)

Digest some data.

pub fn reset(&mut self)

Reset the sum.

pub fn get_crc(&self) -> u8

Get the current CRC value (it always returns a u8 value). You can continue calling digest method even after getting a CRC value.

impl CRCu8

pub fn crc3gsm() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x4	0x3	0x0	false	0x7

let mut crc = CRCu8::crc3gsm();
crc.digest(b"123456789");

pub fn crc4itu() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x7	0x3 (rev: 0xC)	0x0	true	0x0

let mut crc = CRCu8::crc4itu();
crc.digest(b"123456789");

pub fn crc4interlaken() -> CRCu8

Check	Poly	Init	Ref	XorOut

0xB	0x3	0xF	false	0xF

let mut crc = CRCu8::crc4interlaken();
crc.digest(b"123456789");

pub fn crc5epc() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x00	0x09	0x09	false	0x00

let mut crc = CRCu8::crc5epc();
crc.digest(b"123456789");

pub fn crc5itu() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x07	0x15 (rev: 0x15)	0x00	true	0x00

let mut crc = CRCu8::crc5itu();
crc.digest(b"123456789");

pub fn crc5usb() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x19	0x05 (rev: 0x14)	0x1F	true	0x1F

let mut crc = CRCu8::crc5usb();
crc.digest(b"123456789");

pub fn crc6cdma2000_a() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x0D	0x27	0x3F	false	0x00

let mut crc = CRCu8::crc6cdma2000_a();
crc.digest(b"123456789");

pub fn crc6cdma2000_b() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x3B	0x07	0x3F	false	0x00

let mut crc = CRCu8::crc6cdma2000_b();
crc.digest(b"123456789");

pub fn crc6darc() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x26	0x19 (rev: 0x26)	0x00	true	0x00

let mut crc = CRCu8::crc6darc();
crc.digest(b"123456789");

pub fn crc6gsm() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x13	0x2F	0x00	false	0x3F

let mut crc = CRCu8::crc6gsm();
crc.digest(b"123456789");

pub fn crc6itu() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x06	0x03 (rev: 0x30)	0x00	true	0x00

let mut crc = CRCu8::crc6itu();
crc.digest(b"123456789");

pub fn crc7() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x75	0x09	0x00	false	0x00

let mut crc = CRCu8::crc7();
crc.digest(b"123456789");

pub fn crc7umts() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x61	0x45	0x00	false	0x00

let mut crc = CRCu8::crc7umts();
crc.digest(b"123456789");

pub fn crc8() -> CRCu8

Check	Poly	Init	Ref	XorOut

0xF4	0x07	0x00	false	0x00

let mut crc = CRCu8::crc8();
crc.digest(b"123456789");

pub fn crc8cdma2000() -> CRCu8

Check	Poly	Init	Ref	XorOut

0xDA	0x9B	0xFF	false	0x00

let mut crc = CRCu8::crc8cdma2000();
crc.digest(b"123456789");

pub fn crc8darc() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x15	0x39 (rev: 0x9C)	0x00	true	0x00

let mut crc = CRCu8::crc8darc();
crc.digest(b"123456789");

pub fn crc8dvb_s2() -> CRCu8

Check	Poly	Init	Ref	XorOut

0xBC	0xD5	0x00	false	0x00

let mut crc = CRCu8::crc8dvb_s2();
crc.digest(b"123456789");

pub fn crc8ebu() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x97	0x1D (rev: 0xB8)	0xFF	true	0x00

let mut crc = CRCu8::crc8ebu();
crc.digest(b"123456789");

pub fn crc8icode() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x7E	0x1D	0xFD	false	0x00

let mut crc = CRCu8::crc8icode();
crc.digest(b"123456789");

pub fn crc8itu() -> CRCu8

Check	Poly	Init	Ref	XorOut

0xA1	0x07	0x00	false	0x55

let mut crc = CRCu8::crc8itu();
crc.digest(b"123456789");

pub fn crc8maxim() -> CRCu8

Check	Poly	Init	Ref	XorOut

0xA1	0x31 (rev: 0x8C)	0x00	true	0x00

let mut crc = CRCu8::crc8maxim();
crc.digest(b"123456789");

pub fn crc8rohc() -> CRCu8

Check	Poly	Init	Ref	XorOut

0xD0	0x07 (rev: 0xE0)	0xFF	true	0x00

let mut crc = CRCu8::crc8rohc();
crc.digest(b"123456789");

pub fn crc8wcdma() -> CRCu8

Check	Poly	Init	Ref	XorOut

0x25	0x9B (rev: 0xD9)	0x00	true	0x00

let mut crc = CRCu8::crc8wcdma();
crc.digest(b"123456789");